star_date_vs_gameday/requirements.txt
python-dateutil>=2.8.1